requirements.txt
django
python-decouple


django-simple-captcha